pyproject.toml
[tool.poetry]
name = "empire-cli"
version = "0.1.0"
description = ""
authors = ["BC Security <info@bc-security.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
terminaltables = "^3.1.0"
requests = "^2.24.0"
prompt-toolkit = "^3.0.8"
docopt = "^0.6.2"
python-socketio = {extras = ["client"], version = "^4.6.0"}
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
